+++ odra-casper/test-vm/resources/proxy_caller_with_return.wasm
