+++ odra-casper/test-vm/src/vm/mod.rs
@@ -1,0 +1,1 @@
+pub mod casper_vm;
