+++ odra-casper/wasm-env/src/consts.rs
