+++ odra2/Cargo.toml
@@ -12,8 +12,8 @@ odra-core = { path = "../odra-core" }
 odra-types = { path = "../types" }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
-odra-casper-backend2 = { path = "../odra-casper/backend2" }
+odra-casper-wasm-env = { path = "../odra-casper/wasm-env" }
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-odra-casper-test-env2 = { path = "../odra-casper/test-env2" }
+odra-casper-test-vm = { path = "../odra-casper/test-vm" }
 odra-vm = { path = "../odra-vm" }
